fix: accuracy audit round 7 -- celery safety, config hardening, stddev stability
Infrastructure:
- Set Redis visibility_timeout=4200 to prevent duplicate task execution
  with task_acks_late (was below task_time_limit causing re-delivery)
- Add worker_max_tasks_per_child=200 to recycle workers and cap memory
- Add broker_connection_retry_on_startup for Celery 5.3+ compat
- Remove unused market_data queue declaration

Security and config:
- Require CLERK_JWT_KEY or CLERK_JWKS_URL for JWT signature verification
  (CLERK_ISSUER alone previously satisfied the redundant first check)
- Validate log_level against valid Python logging levels
- Add 7 rate-limit count fields to validate_positive_ints
- Constrain db_pool_max_overflow >= 0 (prevent unlimited overflow)
- Change catalog Cache-Control from public to private (authenticated)

Strategy correctness:
- Fix Collar detail_json to use standard field names (identifier,
  share_quantity_per_unit, expiration_date, quantity_per_unit)
- Replace rolling_stddev with Welford's algorithm for numerical stability
- Fix resolve_wing_strike to reject directionally-wrong strikes
- Floor calendar spread capital at 1.0 to prevent zero-capital skip
- Coerce custom strategy quantity_ratio to int at assignment

Scripts and client:
- Anchor check_service_close.py and check_migration_drift.py paths
  to project root instead of CWD-relative
- Add AsyncFunctionDef detection for async close() methods
- Add SyntaxError warning and engine.dispose() to migration drift script
- Fix fetchForecast truthiness check to use != null for horizonDays

Made-with: Cursor

+++ src/backtestforecast/backtests/strategies/custom.py
@@ -82,7 +82,7 @@ class CustomNLegStrategy(StrategyDefinition):
                     OpenStockLeg(
                         symbol=config.symbol,
                         side=side_sign,
-                        share_quantity_per_unit=100 * leg_def.quantity_ratio,
+                        share_quantity_per_unit=int(100 * leg_def.quantity_ratio),
                         entry_price=bar.close_price,
                         last_price=bar.close_price,
                     )
@@ -135,7 +135,7 @@ class CustomNLegStrategy(StrategyDefinition):
                     side=side_sign,
                     strike_price=strike,
                     expiration_date=exp,
-                    quantity_per_unit=leg_def.quantity_ratio,
+                    quantity_per_unit=int(leg_def.quantity_ratio),
                     entry_mid=quote.mid_price,
                     last_mid=quote.mid_price,
                 )
